+++ app/Session.php
@@ -112,8 +112,8 @@ class Session extends Model
         $this->active      = false;
 
         // TODO Cleanup
-        $s = MemorySession::start();
-        $s->set('password', $password);
+        $session = MemorySession::start();
+        $session->set('password', $password);
     }
 
     public function getUploadService()
@@ -146,9 +146,9 @@ class Session extends Model
 
     public function loadMemory()
     {
-        $s = MemorySession::start();
-        $s->set('jid', $this->user_id);
-        $s->set('host', $this->host);
-        $s->set('username', $this->username);
+        $session = MemorySession::start();
+        $session->set('jid', $this->user_id);
+        $session->set('host', $this->host);
+        $session->set('username', $this->username);
     }
 }