+++ lib/moxl/src/Xec/Action/MAM/Get.php
@@ -21,11 +21,11 @@ class Get extends Action
 
     public function request()
     {
-        $sess = Session::start();
+        $session = Session::start();
 
         // Generating the queryid key.
         $this->_queryid = \generateKey(12);
-        $sess->set('mamid'.$this->_queryid, true);
+        $session->set('mamid'.$this->_queryid, true);
         $this->store();
 
         MAM::get(
@@ -49,8 +49,8 @@ class Get extends Action
 
     public function handle($stanza, $parent = false)
     {
-        $sess = Session::start();
-        $sess->remove('mamid'.$this->_queryid);
+        $session = Session::start();
+        $session->delete('mamid'.$this->_queryid);
 
         //MessageBuffer::getInstance()->save();